+++ src/main/java/br/com/pi/projeto_RD/model/dto/SubCategoriaDTO.java
@@ -3,11 +3,12 @@ package br.com.pi.projeto_RD.model.dto;
 import lombok.Data;
 
 import javax.persistence.Column;
+import java.math.BigInteger;
 
 @Data
 public class SubCategoriaDTO {
 
-    private Long idSubCategoria;
+    private BigInteger idSubCategoria;
     private String dsSubCategoria;
 
 }
