+++ src/main/java/br/com/pi/projeto_RD/repository/StatusRepository.java
@@ -4,6 +4,8 @@ import br.com.pi.projeto_RD.model.entity.ProdutoEntity;
 import br.com.pi.projeto_RD.model.entity.StatusProdutoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<StatusProdutoEntity, Long> {
+import java.math.BigInteger;
+
+public interface StatusRepository extends JpaRepository<StatusProdutoEntity, BigInteger> {
 
 }
